Updated to 0.1.8
Updated repo to 0.1.8 based on the changes here https://www.curseforge.com/minecraft/mc-mods/sync-forge
- Added config options to require specific items for shell construction
- Added config options to disable animations when syncing
- Added Curios Compatability
- Fixed some running animation issues
- Fixed some syncing issues
- Fixed some packet issues
- Fixed some issues with gravestone mods
- Fixed some server issues
- Fixed some shell ownership validation during respawn

+++ src/main/java/net/sumik/sync/common/block/entity/ShellStorageBlockEntity.java
@@ -84,7 +84,7 @@ public class ShellStorageBlockEntity extends AbstractShellContainerBlockEntity i
         }
 
         if (!infinitePower && !isReceivingRedstonePower && hasEnergy) {
-            this.storedEnergy = (int) Mth.clamp(this.storedEnergy - config.shellStorageConsumption(), 0, config.shellStorageCapacity());
+            this.storedEnergy = (int) Mth.clamp((float)(this.storedEnergy - config.shellStorageConsumption()), 0.0F, (float)config.shellStorageCapacity());
         }
     }
